requirements.txt
selenium>=4.18.1
pytest>=8.0.2
pytest-html>=4.1.1
pytest-xdist>=3.3.1
allure-pytest>=2.13.2
webdriver-manager>=4.0.1
python-dotenv>=0.21.0
pytest-rerunfailures>=12.0
pytest-timeout>=2.2.0
pytest-selenium>=4.0.0 